REFACTOR setting environment to custom hook
