+++ pwa/static/.env.development
@@ -25,4 +25,4 @@ GATSBY_FOOTER_CONTENT_HEADER=""
 GATSBY_OIDN_NUMBER=""
 
 # Deployment option, if you're using this .env file DO NOT REMOVE OR EDIT THIS
-# GATSBY_ENV_VARS_SET="true"
+GATSBY_ENV_VARS_SET="true"
